requirements.txt
astroid==1.6.3
azure-common==1.1.23
azure-core==1.0.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-queue==12.0.0
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
colorama==0.3.9
cryptography==2.8
grpcio==1.23.0
idna==2.8
isodate==0.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mailchimp3==3.0.10
mccabe==0.6.1
msrest==0.6.10
oauthlib==3.1.0
pycparser==2.19
pylint==1.8.4
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.0
simple-crypt==4.1.7
six==1.11.0
stripe==2.40.0
urllib3==1.25.3
wrapt==1.10.11
